+++ src/model/CommandDrawGrid.cpp
@@ -6,14 +6,15 @@
 #include <include/model/transform.hpp>
 #include <iostream>
 
-CommandDrawGrid::CommandDrawGrid(const QPoint &center, const QPoint &first, const QPen &pen, int number, float opacity) :
-    center(center), first(first), pen(pen), number(number), opacity(opacity) {}
+CommandDrawGrid::CommandDrawGrid(const QPoint &center, const QPoint &first, const QPen &pen, int number, float opacity)
+        :
+        center(center), first(first), pen(pen), number(number), opacity(opacity) {}
 
 void CommandDrawGrid::draw(QPainter &painter) const {
-    int angle = 360 / number;
+    double angle = 360.0 / number;
     if (number > 1) {
         for (int i = 0; i < number; ++i) {
-            QPoint lastPointRotated = rotatePoint(first, center, angle * i);
+            QPoint lastPointRotated = rotatePoint(first, center, angle * (double) i);
             painter.setPen(pen);
             painter.setOpacity(opacity);
             painter.drawLine(center.x(), center.y(), lastPointRotated.x(), lastPointRotated.y());
